feat: post edit

+++ blog-admin/src/hooks/posts.js
@@ -1,11 +1,16 @@
 import {useEffect, useState} from "react";
 import axios from "axios";
+import client from "../client";
+import { useNavigate } from "react-router-dom";
 
 export function usePosts() {
+  const [loading, setLoading] = useState(false);
   const [posts, setPosts] = useState([]);
   const [errorMessage, setErrorMessage] = useState('');
+  const navigate = useNavigate();
 
   const getPosts = async () => {
+    setLoading(true);
     try {
       const response = await axios.get("http://localhost:3004/posts");
       setPosts(response.data);
@@ -13,38 +18,29 @@ export function usePosts() {
       const error = e;
       console.log(error.message);
     }
+    setLoading(false);
   };
 
   const onDeleteHandler = (id) => {
-    axios.delete(`http://localhost:3004/posts/${id}`);
-    window.location.reload();
-  }
-
-  const onSubmitHandler = (e, title, body) => {
-    e.preventDefault();
-    if (title.trim().length > 0 && body.trim().length > 0) {
-      try {
-        axios.post('http://localhost:3004/posts', {
-          title: title,
-          body: body
-      })
-      setErrorMessage('');
-      window.location.pathname = '/';
-      } catch (e) {
-        setErrorMessage(e.message);
-      }
-    } else {
-      setErrorMessage('Invalid input data!');
+    if (window.confirm('Are you really want to delete this post?')) {
+      axios.delete(`http://localhost:3004/posts/${id}`);
+      getPosts();
     }
   }
+  
 
-  const onUpdateHandler = (id, title, body) => {
-    
+  const onSubmitHandler = async ({ title, body }) => {
+    if (title.trim().length > 0 && body.trim().length > 0) {
+      await client.post("/posts", { title, body })
+      navigate(-1)
+    } else {
+      setErrorMessage('Error! You must fill two fields to create new post!')
+    }
   }
 
   useEffect(() => {
     getPosts();
   }, []);
 
-  return { posts, getPosts, onDeleteHandler, onSubmitHandler, onUpdateHandler, errorMessage }
+  return { posts, getPosts, onDeleteHandler, onSubmitHandler, errorMessage, loading }
 }
